Early implementation and test of GUI
Has a working window where the user is prompted for a vector. The user can insert input why typing into the textbox below the prompt, and the user's input will the shown below the textbox as it is being typed. This is just testing, but in the future this will be used to have the app draw vectors as the user type in their input.